Move files to components folder, add startDate and dropdown.
