+++ client/src/app/schedule/components/dropdown.jsx
@@ -1,18 +1,25 @@
-const dropdown = ({dropdownDefault, onDropDownSelect}) => (
-    <div>
+import React from 'react';
+import DropDownMenu from 'material-ui/DropDownMenu';
+import MenuItem from 'material-ui/MenuItem';
+
+
+const DropdownList = (props) => {
+  return (
+    <div className="dropdown">
     <DropDownMenu
       className="dropdown"
       bsStyle="default"
       title="Frequency"
       value={this.props.dropdownValue}
-      onChange={() => onDropDownSelect() /* add in parameter from MenuItem */}
-    >
+      onChange={() => onDropDownSelect() /* add in parameter from MenuItem */
+      }>
       <MenuItem value={1} primaryText="Daily">Daily</MenuItem>
       <MenuItem value={2} primaryText="Weekly">Weekly</MenuItem>
       <MenuItem value={3} primaryText="Monthly">Monthly</MenuItem>
     </DropDownMenu>
     </div>
+  );
+};
 
-);
 
-window.dropdown = dropdown;
+export default DropdownList;
