+++ client/src/app/schedule/components/navBar.jsx
