+++ client/src/app/schedule/components/selectDays.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import DatePicker from 'material-ui/DatePicker';
+
+const SelectDays = () => (
+  <div>
+    <DatePicker hintText="Portrait Dialog" />
+  </div>
+);
+
+export default SelectDays;
