+++ client/src/app/schedule/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ScheduleContainer from './components/scheduleContainer.jsx';
+
+ReactDOM.render(<ScheduleContainer />, document.getElementById('root'));
+
